+++ src/fixpoint/workflows/structured/_context.py
@@ -7,6 +7,7 @@ from fixpoint.agents import BaseAgent
 from fixpoint.cache import SupportsChatCompletionCache
 
 from ..imperative import WorkflowContext as ImperativeWorkflowContext, WorkflowRun
+from ..imperative._wrapped_workflow_agents import WrappedWorkflowAgents
 from ._run_config import RunConfig
 
 
@@ -38,6 +39,8 @@ class WorkflowContext(ImperativeWorkflowContext):
         agents: List[BaseAgent],
         cache: Optional[SupportsChatCompletionCache] = None,
         logger: Optional[logging.Logger] = None,
+        *,
+        _workflow_agents_override_: Optional[WrappedWorkflowAgents] = None,
     ) -> None:
         super().__init__(workflow_run, agents, cache, logger)
         self.run_config = run_config
